fix #1244 simplify Build.scala
